+++ src/framework/graphics/vertex.cpp
@@ -1,6 +1,6 @@
 #include "vertex.h"
 
-Purity::Vertex::Vertex(const Vector2f& position)
+Purity::Vertex::Vertex(const Vector2i& position)
 {
     this->position = position;
 }
